GameManager & Player Movement
Created the game manager including the basics of its script.

Created a prefab for the player. Worked on the player's movement script.

Have not set up inputs yet.

Imported a .png to use as a stand-in sprite for the player.

Created a NoFriction physics material to be used on the tilemap.